Split InstructionRunTest into per-class test files
This is to make the test code more readable.
